test(bedrock_mantle): consolidate the pricing checks in the mapped test file
TestBedrockMantlePricing set LITELLM_LOCAL_MODEL_COST_MAP after import and
never reloaded litellm.model_cost, so it only passed by side effect of the
module-level map and failed against the network map once the rate changed.
It now uses the file's local_cost_map fixture, which reloads the bundled map,
clears the get_model_info cache, and restores everything afterwards.

The offer-file and Converse-parity checks move from a sibling top-level file
into this mapped test file, per the repository test layout rule.
